Implement smallest divisor function with binary search
Example 1:

Input: nums = [1,2,5,9], threshold = 6
Output: 5
Explanation: We can get a sum to 17 (1+2+5+9) if the divisor is 1. 
If the divisor is 4 we can get a sum of 7 (1+1+2+3) and if the divisor is 5 the sum will be 5 (1+1+1+2). 

NAUKARI SAME -> https://www.naukri.com/code360/problems/smallest-divisor-with-the-given-limit_1755882?leftPanelTabValue=PROBLEM

LC -> https://leetcode.com/problems/find-the-smallest-divisor-given-a-threshold/description/

KOKO BANANA -> https://www.naukri.com/code360/problems/koko-eating-bananas_10870969?leftPanelTabValue=PROBLEM